requirements.txt
speedtest-cli>=2.1.3